+++ src/editor/EditorModule.ts
@@ -5,6 +5,7 @@ import { ActivateResult, ExtensionModule } from "../ExtensionModule";
 import { Command } from "../constants";
 import { showErrorMessage } from "../util";
 
+import { DebugSessionRegistry } from "./DebugSessionRegistry";
 import { EditorContentsCommand } from "./EditorContentsCommand";
 
 type EditorContentsCommandHandler = (
@@ -15,10 +16,14 @@ export class EditorModule extends ExtensionModule {
   protected get containerModule(): ContainerModule {
     return new ContainerModule((bind) => {
       bind(EditorContentsCommand).toSelf();
+      bind(DebugSessionRegistry).toSelf().inSingletonScope();
     });
   }
 
   activate(): ActivateResult {
+    const registry = this.container.resolve(DebugSessionRegistry);
+    const disposables = registry.activate();
+
     const runEditorContentsDisposable = vscode.commands.registerCommand(
       Command.runEditorContents,
       this.createHandler()
@@ -29,7 +34,11 @@ export class EditorModule extends ExtensionModule {
       this.createHandler(true)
     );
 
-    return [runEditorContentsDisposable, debugEditorContentsDisposable];
+    return [
+      ...disposables,
+      runEditorContentsDisposable,
+      debugEditorContentsDisposable,
+    ];
   }
 
   private createHandler(debug = false): EditorContentsCommandHandler {